+++ expense-manager-api/src/main/java/com/sivalabs/expensemanager/entities/User.java
@@ -1,16 +1,15 @@
 package com.sivalabs.expensemanager.entities;
 
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import lombok.Setter;
-
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.Table;
 import javax.validation.constraints.NotEmpty;
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
 
 @Entity
 @Table(name = "USERS")
@@ -21,11 +20,9 @@ import javax.validation.constraints.NotEmpty;
 public class User {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
-    @NotEmpty
-    private String name;
-    @NotEmpty
-    private String email;
-    @NotEmpty
-    private String password;
+    private Long id;
+
+    @NotEmpty private String name;
+    @NotEmpty private String email;
+    @NotEmpty private String password;
 }
